+++ source/h1_reach/h1_reach/__init__.py
@@ -3,12 +3,7 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-"""
-Python module serving as a project/extension template.
-"""
+"""OneRobotics A1 external Isaac Lab integration."""
 
 # Register Gym environments.
 from .tasks import *
-
-# Register UI extensions.
-from .ui_extension_example import *